Final push
The app is working! Open docker, and powershell cd to the the project folder, run setup.psi and let it build the containers.. next run the init_db.py with docker exec -it marketproj-backend-1 python init_db.py...  you should get no errors and a message the database was seeded with data. the app is live localhost:3000.