added admin support to force market opan and closed
OK guys this should be the last push. I added a protocol for the admin to force open the market so we can buy and sell on weekends and holidays. There is no visibility to the user meaning if the admin forced open the market, the dashboard still shows closed and normal hours.